replace type header by smaller headers for ptref
